fix: derived new approximation for long times
The approximation we were using for the z part of the greens function at
long times was breaking down in some cases. The problem was that we get
numerical overflow/underflow when mua is large, and the approximation
did not depend on mua.

The new approximatino will allow us to keep from overflowing for a
longer time.