feat: adding optimizations to CLI commands for large batch processing.
Adding optimizations to help with processing large number of
configurations. the `temperature-rise` and `multiple-pulse` sub-commands
now take a --print-ids flag to print the ID for all configurations
defined in the config file(s) and a --ids option to list the config IDs
to run. This allows batch scripts to querry the IDs of the
configurations defined, determine which configurations have already been
ran, and only run those configurations that are missing.