regression: disabling WASM-based Marcum-Q function
We are getting a WASI error when trying to run retin-therm on some batch
configurations. Not sure if there is a limit to the number of wasmer
instances that can be running in different processes, or if we need to
be more careful about how the module get loaed in forked processes.

Needs more investigation...