feat: added support for removing configurations from input set to the multiple-pulse sub command.
With multiple-pulse configurations we need to be careful not to create
invalid configurations. For example, an exposure that lasts longer than
the CW input file covers or trains that have a pulse period that is
shorter than the pulse duration.

Sometimes with batch configurations it is more convienient to just
generate a bunch of configurations and then throw out the ones that are
invalid, rather than trying to only create valid configurations.

This is what the optional '/remove' key is for. If given, it is a list
of bool values. If any value in the list is True, the configuration will
be removed.

With powerconf, the values can be computed dynamically from other
configuration parameters. i.e.

remove:
  - $( ${tau} > ${t0} )  # remove configurations that have a DC > 100%
  - $( ${tmax}.to("s").magnitude > 100 ) # remove any exposures that are
longer than 100s.