devel: working on schemas that can be used for validating cli subcommand
Trying to find the best way to validate input configurations for the cli
sub-commands. I.e. `retina-therm temperature-rise` and `retina-therm
multiple-pulse`. Don't really want to write logic by hand, but we are
using an fspathtree to handle configuration since we are using
powerconf.

Maybe we need to render th econfigs with powerconf and then pass it
throuh pydantic...